refactor: rename restart-game-buttom to restart-game-button (NAME-01)
