+++ src/app/main-game/game-over/game-over.component.html
@@ -60,8 +60,8 @@
             </div>
         </div>
     </div>
-    <app-restart-game-buttom
+    <app-restart-game-button
         (restartEvent)="resetGameAction()">
-    </app-restart-game-buttom>
+    </app-restart-game-button>
 </div>
