+++ src/app/main-game/game-over/game-over.component.ts
@@ -14,7 +14,7 @@ import { GymLeader } from '../../interfaces/gym-leader';
 import { gymLeadersByGeneration } from '../roulette-container/roulettes/gym-battle-roulette/gym-leaders-by-generation'; 
 import { eliteFourByGeneration } from '../roulette-container/roulettes/elite-four-battle-roulette/elite-four-by-generation';
 import { championByGeneration } from '../roulette-container/roulettes/champion-battle-roulette/champion-by-generation';
-import { RestartGameButtonComponent } from "../../restart-game-buttom/restart-game-buttom.component";
+import { RestartGameButtonComponent } from "../../restart-game-button/restart-game-button.component";
 import {TranslatePipe} from '@ngx-translate/core';
 @Component({
   selector: 'app-game-over',
