+++ src/app/main-game/main-game.component.html
@@ -9,7 +9,7 @@
                     <app-coffee-button></app-coffee-button>
                 </div>
                 <div class="col-6 col-sm-12 mb-2">
-                    <app-restart-game-buttom (restartEvent)="resetGameAction()"></app-restart-game-buttom>
+                    <app-restart-game-button (restartEvent)="resetGameAction()"></app-restart-game-button>
                 </div>
                 <div class="col-6 col-sm-12 mb-2">
                     <app-language-selector></app-language-selector>
